backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
torch==2.5.1
torchvision==0.20.1
efficientnet-pytorch==0.7.1
Pillow==11.1.0
numpy==2.1.3
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
openai==1.59.9
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
bcrypt==4.2.1
aiofiles==24.1.0
